fixing bugs with google login

+++ routes/api.php
@@ -17,6 +17,14 @@ use App\Http\Middleware\CheckManagerRole;
 use App\Http\Controllers\ContactController;
 use App\Models\User;
 use App\Http\Controllers\PersonController;
+use App\Http\Controllers\AuthController;
+use Laravel\Socialite\Facades\Socialite;
+
+
+
+
+
+
 
 Route::post('login', [LoginController::class, 'login']);
 
@@ -56,47 +64,47 @@ Route::middleware('auth:sanctum')->group(function () {
 Route::middleware(['Manager'])->group(function () {
     
 
-    Route::post('/users/createmanager', [UserController::class, 'createManager']);
+    Route::middleware('auth:sanctum')->post('/users/createmanager', [UserController::class, 'createManager']);
 });
 
 
 
 Route::prefix('itemsupplier')->middleware('Manager')->group(function () {
-    Route::get('/', [ItemSupplierController::class, 'index']);
-    Route::post('/addsupply', [ItemSupplierController::class, 'store']);
-    Route::put('/edit/{id}', [ItemSupplierController::class, 'update']);
-    Route::delete('/delete/{id}', [ItemSupplierController::class, 'destroy']);
+    Route::middleware('auth:sanctum')->get('/', [ItemSupplierController::class, 'index']);
+    Route::middleware('auth:sanctum')->post('/addsupply', [ItemSupplierController::class, 'store']);
+    Route::middleware('auth:sanctum')->put('/edit/{id}', [ItemSupplierController::class, 'update']);
+    Route::middleware('auth:sanctum')->delete('/delete/{id}', [ItemSupplierController::class, 'destroy']);
 });
 
 
 Route::prefix('suppliers')->middleware(['auth:sanctum', 'Manager'])->group(function () {
-    Route::get('/', [SupplierController::class, 'index']);
-    Route::post('/addsupplier', [SupplierController::class, 'store']);
-    Route::put('/edit/{id}', [SupplierController::class, 'update']);
-    Route::delete('/delete/{id}', [SupplierController::class, 'destroy']);
+    Route::middleware('auth:sanctum')->get('/', [SupplierController::class, 'index']);
+    Route::middleware('auth:sanctum')->post('/addsupplier', [SupplierController::class, 'store']);
+    Route::middleware('auth:sanctum')->put('/edit/{id}', [SupplierController::class, 'update']);
+    Route::middleware('auth:sanctum')->delete('/delete/{id}', [SupplierController::class, 'destroy']);
 });
 
 Route::prefix('items')->middleware('Manager')->group(function () {
-    Route::get('/', [ItemController::class, 'index']);
-    Route::post('/addnewitem', [ItemController::class, 'store']);
-    Route::put('/edit/{id}', [ItemController::class, 'update']);
-    Route::delete('/delete/{id}', [ItemController::class, 'destroy']);
+    Route::middleware('auth:sanctum')->get('/', [ItemController::class, 'index']);
+    Route::middleware('auth:sanctum')->post('/addnewitem', [ItemController::class, 'store']);
+    Route::middleware('auth:sanctum')->put('/edit/{id}', [ItemController::class, 'update']);
+    Route::middleware('auth:sanctum')->delete('/delete/{id}', [ItemController::class, 'destroy']);
 
 });
 
 Route::prefix('orders')->middleware('auth:sanctum')->group(function () { //my orders
-    Route::get('/myorders', [OrderController::class, 'MyOrdersindex']);
-    Route::get('/myorder/details/{orderId}', [ItemOrderController::class, 'MyOrderDetails']);
-    Route::post('/addorder', [OrderController::class, 'saveOrder']);
-    Route::put('/update/{id}', [ItemOrderController::class, 'update']);
-    Route::delete('/delete/{id}', [OrderController::class, 'destroy']);
+    Route::middleware('auth:sanctum')->get('/myorders', [OrderController::class, 'MyOrdersindex']);
+    Route::middleware('auth:sanctum')->get('/myorder/details/{orderId}', [ItemOrderController::class, 'MyOrderDetails']);
+    Route::middleware('auth:sanctum')->post('/addorder', [OrderController::class, 'saveOrder']);
+    Route::middleware('auth:sanctum')->put('/update/{id}', [ItemOrderController::class, 'update']);
+    Route::middleware('auth:sanctum')->delete('/delete/{id}', [OrderController::class, 'destroy']);
 });
 
 Route::prefix('orders')->middleware('Manager')->group(function () { //users orders
     
-    Route::get('/', [OrderController::class, 'index']);
-    Route::get('/userorder/details/{orderId}', [ItemOrderController::class, 'OrderDetails']);
-    Route::put('/userorder/update/{id}', [OrderController::class, 'updatePending']);
+    Route::middleware('auth:sanctum')->get('/', [OrderController::class, 'index']);
+    Route::middleware('auth:sanctum')->get('/userorder/details/{orderId}', [ItemOrderController::class, 'OrderDetails']);
+    Route::middleware('auth:sanctum')->put('/userorder/update/{id}', [OrderController::class, 'updatePending']);
 
 
 });
@@ -104,14 +112,14 @@ Route::prefix('orders')->middleware('Manager')->group(function () { //users orde
 Route::middleware('auth:sanctum')->get('/categories', [CategoryController::class, 'index']);
 Route::prefix('categories')->middleware('Manager')->group(function () {
     
-    Route::post('/addcategory', [CategoryController::class, 'store'])->name('categories.store');
-    Route::put('/edit/{id}', [CategoryController::class, 'update']);
-    Route::delete('/delete/{id}', [CategoryController::class, 'destroy']);
+    Route::middleware('auth:sanctum')->post('/addcategory', [CategoryController::class, 'store'])->name('categories.store');
+    Route::middleware('auth:sanctum')->put('/edit/{id}', [CategoryController::class, 'update']);
+    Route::middleware('auth:sanctum')->delete('/delete/{id}', [CategoryController::class, 'destroy']);
 
 });
 
 Route::prefix('users')->middleware('Manager')->group(function () {
-    Route::delete('/{id}', [UserController::class, 'destroy']);
+    Route::middleware('auth:sanctum')->delete('/{id}', [UserController::class, 'destroy']);
 
 
 
@@ -136,5 +144,7 @@ Route::prefix('person')->middleware('auth:sanctum')->group(function () {
     Route::put('/updatepassword', [PersonController::class, 'updateUserPassword']);
     Route::delete('/delete', [PersonController::class, 'deleteUser']);
 });
+
+
